Remove task dependencies when a task exits
When a task exits we have to remove all references to it. It's simpler to
explain with one example:

Suppose we have the following situation, where 'A -> B' means A depends on B,
i.e. B is on A's taskaffinity_list and A is on B's followme_list:

    A ------> B
      \       |
       \      v
        ----> C

Overall situtation:

                    _________               _________       _________
Task:               |   A   |               |   B   |       |   C   |
                    ---------               ---------       ---------
taskaffinity_list:  |   *<--|-->B<--->C     |   *<--|-->C   |   *<--|-->NULL
                    |       |               |       |       |       |
followme_list:      |   *<--|-->NULL        |   *<--|-->A   |   *<--|-->A<---->B
                    ---------               ---------       ---------

If B exits, it means to:
	1) Remove B from the followme_list of A and then remove C from B's
	taskaffinity_list

	2) Remove A from the followme_list of B and then remove B from A's
	taskaffinity_list

This is accomplished by calling sched_del_taskaffinity on all tasks of B's
taskaffinity_list and after calling sched_del_taskaffinity on all tasks of its
followme_list (they will first remove themselves from B's followme_list and
after from their taskaffinity_list). This approach preserves the order
mentioned above.

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>