Updating allowed_cpus according to task-affinity
Create the mechanism to allow a task to follow other tasks. Previous commits
permitted the creation of group of tasks, or better, to create a graph with
tasks' dependency.

This patch, instead, is about checking this graph in the moment a task has
just woken up to decide in which runqueue to put it. The runqueue will be one
of the CPU's runqueue its dependencies were previously running. Example

Suppose we have the following dependencies among tasks:

T1 <--
      \
       \
        T3
       /
      /
T2 <--

This means task T3 depends upon T1 and T2 and will allways run either on the
CPU T1 ran last time or on that where T2 ran.

This is mostly intended for collaborative tasks and the simplest example one
can think is the producer-consumer. In previous example, T1 and T2 are
producers and T3 consumes what T1 and T2 put in a shared buffer. Being on the
same CPU of at least one of them allows T3 to consume data much faster since
it will probably get cache-hot data.

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>