Added task_affinity_rq structure and initialized it
With this struct also runqueues are involved in taskaffinity
This struct has field last_task to store last task executed on
rq. This information is used by consumer to know which cpus
have executed ,as last task, a producer. In this way, consumer
knows that maybe those cpus have still producer's data in their
L1 cache.

Signed-off-by: Matteo Marini <chuckberry26@gmail.com>